feat: customtkinter ui
