+++ audiblez/__main__.py
@@ -1,0 +1,4 @@
+from audiblez.ui import main
+
+if __name__ == "__main__":
+    main()
