+++ audiblez/ctk/constants.py
@@ -1,0 +1,25 @@
+import customtkinter as ctk
+
+# UI Constants
+APP_NAME = "Audiblez"
+DEFAULT_WIDTH = 1200
+DEFAULT_HEIGHT = 800
+
+# Colors (Dark Mode Only)
+BG_COLOR = "#1e1e1e"
+PANEL_COLOR = "#252526"
+BORDER_COLOR = "#3e3e42"
+TEXT_COLOR = "#cccccc"
+HIGHLIGHT_COLOR = "#007acc"
+
+# Synthesis Constants
+PREVIEW_LIMIT = 300
+DEFAULT_SPEED = 1.0
+DEFAULT_ENGINE = "cpu"
+# M4B Methods
+M4B_ORIGINAL = "original"
+M4B_CRISPY = "extra crispy"
+
+# Events (Custom Strings or similar mechanism)
+EVENT_PROGRESS = "progress"
+EVENT_FINISHED = "finished"
