+++ plans/blueprint.md
@@ -15,7 +15,7 @@ The application follows a modular "Engine" pattern, allowing different TTS backe
 1. **Extraction**: `calibre_handler.py` or `core.py` extracts text from EPUB chapters.
 2. **Persistence**: `database.py` manages the state of books being processed (staging) and configuration.
 3. **Processing**: `core.py` coordinates the generation loop, passing text chunks to the selected engine.
-4. **Synthesis**: Engines in `audiblez/engines/` interact with AI models (Kokoro, Qwen3) to produce audio segments.
+4. **Synthesis**: Engines in `audiblez/engines/` interact with AI models (Kokoro now, Qwen3 later) to produce audio segments.
 5. **Assembly**: `ffmpeg` is used to concatenate segments into chapters and finally into a metadata-rich `.m4b` file.
 
 ## File Responsibilities (`audiblez/`)
@@ -32,7 +32,7 @@ The application follows a modular "Engine" pattern, allowing different TTS backe
 - **[engines/](file:///s:/Files/nexus/http/audiblez/audiblez/engines/)**: Modular TTS drivers.
   - `base.py`: Abstract base class for all engines.
   - `kokoro_engine.py`: Default high-speed engine.
-  - `qwen3_engine.py`: Support for advanced LLM-based TTS.
+  - `qwen3_engine.py`: Support for advanced LLM-based TTS. [TODO]
 - **[voices.py](file:///s:/Files/nexus/http/audiblez/audiblez/voices.py)**: Configuration and list of supported TTS voices across different languages.
 
 ### Utilities
@@ -43,7 +43,7 @@ The application follows a modular "Engine" pattern, allowing different TTS backe
 
 ## Development Reference
 
-### Environment
+### Environment [under construction]
 
 - **Python Version**: 3.12 (Strictly recommended for compatibility with specific Windows wheels).
 - **Package Manager**: `uv` is preferred.
